refactor(calendar): extract event CRUD into CalendarEventService
Move single + recurring event CRUD, transfer, instance/expansion reads, and the
event recurrence exception/bulk-modification methods out of the CalendarService
facade into a new CalendarEventService. The facade constructs it per-call from a
live context snapshot (cheap dataclass build, no re-auth) and delegates each
public event method to it.

Availability (Phase 4), bundle-event fan-out (Phase 3), and the shared
write-adapter / attendee-permission helpers stay on the facade and are reached
through the EventServiceHost seam (the facade supplies itself), preserving
behavior byte-for-byte.
